add support for TSInput-style parsing
Extend the Parser.parse() method to take a callable which acts like the
"read" callback from a TSInput struct. This callable takes the requested
byte offset and source location and returns a bytes object corresponding
to the source at that location. Returning an empty bytes object or None
terminates parsing for that line.

It is assumed that the parser only needs to "borrow" the return value
from the TSInput's read callback until the next call to the function or
when the ts_parser_parse() function returns. As such we use the
"payload" of the TSInput structure to record the last bytes object
returned so that its reference count may be decremented on the next call
to the read callback or when ts_parser_parse() returns.

Add some basic tests which demonstrate using the byte offset and TSPoint
arguments.