expose ts_tree_get_changed_ranges() functionality
Expose the ts_tree_get_changed_ranges() functionality by exposing a
method for Tree called get_changed_ranges(). This should be called on
the *old* tree passing in the *new* one. (This mirrors the argument
ordering of ts_tree_get_changed_ranges().)

The return value is a list of Range objects which wrap the TSRange
structure and provide getters for the {start,end}_{point,byte} members.

A simple test is added as part of the Tree test suite and the README is
updated to document the new method.