expose ts_query_cursor_set_{byte,point}_range
Some of the ground work for this had already been done; the
Query.captures() method already had {start,end}_point keyword arguments
but they were ignored.

Complete the keyword arguments by adding {start,end}_byte arguments.
Parse the arguments as Py_ssize_t-s as they need to be large enough to
represent memory offsets. Using a signed type allows the default values
to be -1, in which case the full node range is used which preserves the
original functionality.

Add tests of the byte and point range keywords and update README.