fix: make predicate with a missing capture return True
If a predicate in a query references a capture that was not matched
because it was optional in the query, return true rather than throw an
error. In other words, if the capture was not found, ignore the
predicate entirely.

This matches the behavior of tree-sitter CLI and the playground. It
 is not a common case but we have run into the issue matching Kotlin functions
where the return type of a function is optional in the grammar requiring
us to do additional filtering in the app code after the query.