feat: make match captures a dictionary (#165)
Return the captures of a match as a dict where the key is the capture
name and the value is a Node or a list of Nodes. For regular captures,
the value is a Node. For captures that can contain multiple nodes
because they use a * or + quantifier, the value is a list of Nodes.