fix: pattern_index parameter to generic predicates is incorrect
When a generic predicate callable is used, the documented signature
indicates that the third argument should be the pattern index.
Unfortunately, the value that was being passed was an index into an
internal array of only the queries containing predicates, not the true
index of the pattern. This made it impossible for a custom directive to
store information into the relevant `pattern_settings` dict (similar to
how the built-in `#set!` directive does).

This change fixes that by storing the pattern index into the
`QueryPredicateGeneric` structure, and then plumbing that through to the
predicate callable.